Se refactoriza modelos en el modulo de authentication
