+++ src/app/modules/authentication/models/index.ts
@@ -1,1 +1,3 @@
-export * from './auth.model';
+export * from './login-credentials';
+export * from './usuario-autenticado';
+export * from './usuario';
