+++ src/app/modules/authentication/models/login-credentials.ts
@@ -1,0 +1,4 @@
+export interface LoginCredentials {
+    identificacion: string;
+    password: string;
+}
