+++ src/app/modules/authentication/models/usuario-autenticado.ts
@@ -1,0 +1,6 @@
+export interface UsuarioAutenticado {
+    token: string;
+    bearer: number;
+    identificacion: string;
+    authorities: string[];
+}
