+++ src/app/modules/authentication/models/usuario.ts
@@ -1,0 +1,12 @@
+export interface Usuario {
+    idUsuario?: number;
+    identificacion: string;
+    nombres: string;
+    apellidos: string;
+    email: string;
+    password?: string;
+    telefono: string;
+    direccion: string;
+    ciudad: string;
+    roles: string[];
+}
